Modularize and add CLI options
